refactor (Movie): 액티비티별로 패키지 분리
