+++ app/src/main/java/woowacourse/movie/MainActivity.kt
@@ -3,6 +3,7 @@ package woowacourse.movie
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.movieList.MovieListActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
